Correct type of icon_span
The `icon_span` is used as a `FT_SpanFunc`. To avoid undefined behavior
and a CFI sanitizer report, this function needs to called through a
pointer of the same type. Fix this by making `icon_span` take the right
parameter types and cast the argument in the function body.

* src/ftcommon.c (icon_span): take correct parameters and cast in
function body. (FTDemo_Icon): remove cast