ftlint.1: Instruct man(1) to run the groff `tbl` preprocessor
The latest version of groff (1.23.0) emits a warning when it detects the
use of table macros (.TS, .TE) with no .TW (table width) macro defined:

  tbl preprocessor failed, or it or soelim was not run; table(s) likely
  not rendered (TE macro called with TW register undefined).

The warning occurs because the `tbl` preprocessor was not invoked by
the `man` command. To fix the issue, we need to instruct man(1) to
run the `tbl` preprocessor before rendering the source as text.

This issue was detected by Debian's Lintian tool.